Verified branch

+++ app/Http/Controllers/ChoiceController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Choice;
 use Illuminate\Http\Request;
+use App\Http\Resources\ChoiceResource;
+
 
 class ChoiceController extends Controller
 {
@@ -13,10 +15,10 @@ class ChoiceController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        //
+    {    //
+        $this->authorize('viewAny', Choice::class);
         $questions=Question::all();
-        return view('choice.index', compact('question'));
+        return response(['choices'=>ChoiceResource::collection($choices),'message'=>'Retrieved successfully']);
     }
 
     /**
@@ -27,6 +29,7 @@ class ChoiceController extends Controller
     public function create()
     {
         //
+        $this->authorize('create', Choice::class);
         $question=Question::all();
         return view('choice.create', compact('question'));
     }
@@ -40,7 +43,7 @@ class ChoiceController extends Controller
     public function store(Request $request)
     {
         //
-       
+        $this->authorize('create', Choice::class);
         $input=$request->all();
         $resources=Resource::create($input);
         return redirect('/choice/'.$choice->id);
@@ -55,6 +58,7 @@ class ChoiceController extends Controller
     public function show(Choice $choice)
     {
         //
+        $this->authorize('view', Choice::class);
         $question=Question::all();
         return view('choice.show', compact('question'));
     }
@@ -68,6 +72,7 @@ class ChoiceController extends Controller
     public function edit(Choice $choice)
     {
         //
+        $this->authorize('update', Choice::class);
         $question=Question::all();
         return view('choice.edit',compact('question'));
     }
@@ -82,7 +87,7 @@ class ChoiceController extends Controller
     public function update(Request $request, Choice $choice)
     {
         //
-       
+        $this->authorize('update', Choice::class);
         $input=$request->all();
         $choice->update($input);
         return redirect('choice/'.$choices->id);
@@ -97,6 +102,7 @@ class ChoiceController extends Controller
     public function destroy(Choice $choice)
     {
         //
+        $this->authorize('delete', Choice::class);
         $choice->delete();
         return redirect('/choice');
     }
