+++ app/Http/Controllers/ModuleController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Module;
+use App\Models\Course;
 use Illuminate\Http\Request;
+use App\Http\Resources\ModuleResource;
 
 class ModuleController extends Controller
+
 {
     /**
      * Display a listing of the resource.
@@ -15,8 +18,9 @@ class ModuleController extends Controller
     public function index()
     {
         //
-        $course=Course::all();
-        return view('module.index'compact('course'));
+        $this->authorize('viewAny', Module::class);
+        $modules=Module::all();
+        return response(['user' => UserResource::collection($user), 'message' => 'Retrieved successfully']);
     }
 
     /**
@@ -27,8 +31,10 @@ class ModuleController extends Controller
     public function create()
     {
         //
-        $course=Course::all();
-        return view('module.create',compact('course'));
+        $this->authorize('create', Module::class);
+        $courses=Course::all();
+        $modules=Module::all();
+        return view('module.create',compact('courses','modules'));
     }
 
     /**
@@ -39,8 +45,19 @@ class ModuleController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        //'course_id','title','description']
+        $this->authorize('create', Module::class);
         $input=$request->all();
+         $module=Module::create($input);
+        return redirect('/module/'.$module->id);
+        $validator = Validator::make($input,[
+            'course_id' => 'required',
+            'title' => 'required',
+            'description' => 'required',
+        ]);
+                if($validator->fails()){
+                    return response(['error'=>$validator->errors(), 'Validation Error']);
+                }
         $module=Module::create($input);
         return redirect('/module/'.$module->id);
     }
@@ -54,6 +71,7 @@ class ModuleController extends Controller
     public function show(Module $module)
     {
         //
+        $this->authorize('view', Module::class);
         $course=Course::all();
         return view('module.show', compact('course'));
     }
@@ -67,7 +85,8 @@ class ModuleController extends Controller
     public function edit(Module $module)
     {
         //
-        $course=Course::all();
+        $this->authorize('update', Module::class);
+        $module=Module::all();
         return view('module.edit', compact('course'));
     }
 
@@ -81,6 +100,7 @@ class ModuleController extends Controller
     public function update(Request $request, Module $module)
     {
         //
+        $this->authorize('update', Module::class);
         $input=$request->all();
         $module->update($input);
         return redirect('module/'.$module->id);
@@ -95,6 +115,7 @@ class ModuleController extends Controller
     public function destroy(Module $module)
     {
         //
+        $this->authorize('delete', Module::class);
         $module->delete();
         return redirect('/module');
     }
