+++ app/Http/Controllers/QuestionController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Question;
+use App\Models\Quiz;
 use Illuminate\Http\Request;
+use App\Http\Resources\QuestionResource;
 
 class QuestionController extends Controller
 {
@@ -14,8 +16,11 @@ class QuestionController extends Controller
      */
     public function index()
     {
-        //
-        return view('question.index');
+        //       
+        $this->authorize('viewAny', Question::class);
+        $questions=Question::all();
+        $quizzes=Quiz::all();
+        return response(['question'=>QuestionResource::collection($question), 'message'=>'Retrieved successfully']);
     }
 
     /**
@@ -26,7 +31,10 @@ class QuestionController extends Controller
     public function create()
     {
         //
-        return view('question.create');
+        $this->authorize('create', Question::class);
+        // dd('here');
+         $quizzes=Quiz::all();
+        return view('question.create',compact('quizzes'));
     }
 
     /**
@@ -37,9 +45,20 @@ class QuestionController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        //quiz_id','student','type','answer
+        $this->authorize('create', Question::class);
         $input=$request->all();
         $question=Question::create($input);
+        $validator=Validator::make($input,[
+            'quiz_id'=>'required',
+            'student'=>'required',
+            'type'=>'required',
+            'answer'=>'required',
+        ]);
+                if($validator->fails()){
+                    return response(['error'=>$validator->errors(), 'Validator Error']);
+                }
+        return redirect('/questions'.$questions->id);
     }
 
     /**
@@ -51,6 +70,7 @@ class QuestionController extends Controller
     public function show(Question $question)
     {
         //
+        $this->authorize('view', Question::class);
         return view('question.show');
     }
 
@@ -63,6 +83,7 @@ class QuestionController extends Controller
     public function edit(Question $question)
     {
         //
+        $this->authorize('update', Question::class);
         return view('question.edit');
     }
 
@@ -76,6 +97,7 @@ class QuestionController extends Controller
     public function update(Request $request, Question $question)
     {
         //
+        $this->authorize('update', Question::class);
         $input=$request->all();
         $question->update($input);
     }
@@ -89,6 +111,7 @@ class QuestionController extends Controller
     public function destroy(Question $question)
     {
         //
+        $this->authorize('delete', Question::class);
         $question->delete();
         return redirect('/question');
     }
