+++ app/Http/Controllers/ResourceController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Resource;
+use App\Models\Page;
 use Illuminate\Http\Request;
+use App\Http\Resources\ResourceResource;
 
 class ResourceController extends Controller
 {
@@ -14,10 +16,12 @@ class ResourceController extends Controller
      */
     public function index()
     {
-        //
-        $page=Page::all();
-        $resource=Resource::all();
-        return view('resource.index',compact('page','resource'));
+        //        return response(['user' => UserResource::collection($user), 'message' => 'Retrieved successfully']);
+
+        $this->authorize('viewAny',Resource::class);
+        $pages=Page::all();
+        $resources = Resource::all();
+        return response(['resource'=>ResourceResource::collection($resource), 'message'=>'Retrieved successfully']);
     }
 
     /**
@@ -28,6 +32,7 @@ class ResourceController extends Controller
     public function create()
     {
         //
+        $this->authorize('create',Resource::class);
         $page=Page::all();
         $resource=Resource::all();
         return view('resource.create',compact('page','resource'));
@@ -41,9 +46,20 @@ class ResourceController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        //page_id','type','title','description','url
+        $this->authorize('create',Resource::class);
         $input=$request->all();
         $resource=Resource::create($input);
+        $validator=Validator::make($input,[
+            'page_id'=>'required',
+            'type'=>'required',
+            'title'=>'required',
+            'description'=>'required',
+            'url'=>'required',
+        ]);
+        if($validator->fails()){
+            return response(['error'=>$validator->errors(), 'Validator Error']);
+        }
         return redirect('/resource/'.$resource->id);
     }
 
@@ -56,6 +72,7 @@ class ResourceController extends Controller
     public function show(Resource $resource)
     {
         //
+        $this->authorize('view',Resource::class);
         $page=Page::all();
         $resource=Resource::all();
         return view('resource.show',compact('page','resource'));
@@ -70,6 +87,7 @@ class ResourceController extends Controller
     public function edit(Resource $resource)
     {
         //
+        $this->authorize('update',Resource::class);
         $page=Page::all();
         return view('resource.edit',compact('page','resource'));
     }
@@ -84,6 +102,7 @@ class ResourceController extends Controller
     public function update(Request $request, Resource $resource)
     {
         //
+        $this->authorize('update',Resource::class);
         $input=$request->all();
         $resource->update('$input');
         return redirect('resource/'.$resource->id);
@@ -98,6 +117,7 @@ class ResourceController extends Controller
     public function destroy(Resource $resource)
     {
         //
+        $this->authorize('delete',Resource::class);
         $resource->delete();
         return redirect('/resource');
     }
